refactor: 일최고, 최저온도 타입 int -> double

+++ backend/domain-modules/forecast-domain/src/main/java/com/softeer/entity/ShortForecastDetailEntity.java
@@ -32,7 +32,7 @@ public class ShortForecastDetailEntity {
     private double lowestTemperature;
 
     public ShortForecastDetailEntity(ForecastEntity forecast, double precipitationProbability,  double snowAccumulation,
-                                     int highestTemperature, int lowestTemperature
+                                     double highestTemperature, double lowestTemperature
                                      ) {
         this.forecast = forecast;
         this.precipitationProbability = precipitationProbability;
